修改Android消息通知的延迟时间 修改加速度计的更新间隔

+++ CrossApp/platform/CAAccelerometerDelegate.h
@@ -16,7 +16,7 @@ public:
     double y;
     double z;
 
-    double timestamp;
+    float timestamp;
 };
 
 /**
